texcomp/texpipe: BC6H float decode path (tp_ktx2_decode_level_rgbaf)
BC6H was the last codec the KTX2 reader could not decode, and it could not be
fixed on the RGBA8 path: it is HDR, so 8-bit is the wrong target.

texcomp gains tc_bc6h_decompress_rgb16f (FP16, BC6H's natural output) and
tc_bc6h_decompress_rgbaf (float RGBA, alpha 1 -- BC6H carries no alpha).
`is_signed` selects sf16 over uf16; the two disagree on endpoint unquantisation,
so it has to be right.

The decoder is the bcdec port that already lived in test/bc6h_ref_decode.h as
the oracle for texcomp-bc6h-gate. It moves to src/texcomp_bc6h_decode.c, so
there is one implementation rather than two copies that can drift; the test
header becomes a shim over the public API and both BC6H gates keep working
unchanged. Their purpose is unchanged too: they still validate the in-house
BC6H encoder against a decoder of independent lineage. The gate reports exactly
the same PSNR as before the move (50.07 / 34.90 dB), which pins the behaviour.

texpipe gains tp_ktx2_decode_level_rgbaf / tp_ktx2_decode_slice_rgbaf. The
slice-locating and bounds-checking logic is now a shared helper, so the float
and 8-bit paths cannot drift in their overflow checks (the surface size is
computed in 64 bits for both). tp_ktx2_image grows is_signed, mapped from the
VkFormat (BC6H_UFLOAT 143 vs BC6H_SFLOAT 144).

The LDR codecs deliberately stay on the RGBA8 path rather than being widened to
float, and ASTC HDR still has no decoder; both return TP_ERROR_UNSUPPORTED from
the float path, and BC6H returns TP_ERROR_UNSUPPORTED from the 8-bit one.

Tests: BC6H KTX2 write -> read -> float decode for both uf16 and sf16 (41.2 /
39.8 dB), alpha == 1.0, short-output rejection, and the rgba8 path refusing
BC6H. Verified non-vacuous: hard-coding is_signed = 0 drops sf16 to 7.9 dB and
fails the floor.
